requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
colorama>=0.4.6
requests>=2.31.0
websocket-client>=1.6.0
pytest>=7.4.0
tabulate>=0.9.0
